Rearrange tests for the daily digest
